Cease logging usertime
Was logging usertime. But, PLRM3 p714: “usertime … returns the value of a clock that is incremented by 1 for every millisecond of execution by the PostScript interpreter. The value has no defined meaning in terms of calendar time or time of day; its only use is interval timing.” So no longer logging usertime.