Inconsistent randomisation fixed
Bug with randomisation, the sequences being the same except when not (which is rare). Maybe a problem with, say, being near a rounding boundary causing more invocations of rand. Whatever the problem, not properly found, but nonetheless seemingly fixed by computing the starting value, rather than it merely being the next at the end of the previous circle’s.