DropletsCharges, location can be array of integers
`DropletsCharges`, location can be array of integers, being the average of the centres of these WithinPage circles. This is a change of meaning for locations that were a length-two array of integers: was [x y], now [WithinPage0 WithinPage1]. It is believed that there were in the wild no instances of this.